Auto 2026-06-26 15:03

+++ src/u2/08fnCASE.sql
@@ -1,5 +1,5 @@
 delimiter $$
-create function fnCase (dato double)
+create function fnCase (dato int)
 returns char(10) deterministic
 begin
    declare leyenda char(10);
